create variable_utils
