+++ paddle/fluid/framework/variable_utils.h
@@ -1,0 +1,34 @@
+//  Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#pragma once
+
+#include "paddle/fluid/framework/variable.h"
+
+#include "paddle/phi/common/int_array.h"
+#include "paddle/phi/common/scalar.h"
+
+namespace paddle {
+namespace experimental {
+
+phi::Scalar MakePhiScalarFromVar(const framework::Variable& variable);
+
+phi::IntArray MakePhiIntArray(const phi::DenseTensor& src);
+
+phi::IntArray MakePhiIntArrayFromVar(const framework::Variable& variable);
+
+phi::IntArray MakePhiIntArrayFromVarList(
+    const std::vector<framework::Variable*>& variable_list);
+
+}  // namespace experimental
+}  // namespace paddle
