+++ paddle/phi/api/lib/utils/CMakeLists.txt
@@ -1,13 +1,4 @@
 cc_library(
   phi_api_utils
   SRCS tensor_utils.cc
-  DEPS tensor_base
-       convert_utils
-       dense_tensor
-       lod_tensor
-       selected_rows_utils
-       place
-       var_type_traits
-       string_tensor
-       int_array
-       scalar)
+  DEPS dense_tensor)
